requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
certifi==2024.8.30
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.2
Flask==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
MarkupSafe==2.1.5
openai==1.44.0
psycopg2-binary==2.9.9
pydantic==2.9.0
pydantic_core==2.23.2
python-dotenv==1.0.1
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
Werkzeug==3.0.4
